Fix menu mobile version (#55)
* Implemeted menu for mobile

* Remove id avatar

* Improve responsive

* Separate UserInfo from the header and refactor structuring

* Fix the avatar to the same line of the dropdown

* Remove unused import
