+++ apps/admin_panel/assets/src/pages/authenticated/AuthenticatedLayout.jsx
@@ -3,17 +3,17 @@ import { connect } from 'react-redux';
 
 import PropTypes from 'prop-types';
 import Alerter from '../../components/Alerter';
-import Header from './Header';
-import Sidebar from './Sidebar';
+import Header from './header/Header';
+import Sidebar from './sidebar/Sidebar';
 
 const AuthenticatedLayout = ({ alert, children }) => (
   <div className="fh">
     <Header />
     <div className="fh">
-      <div className="col-md-3 authenticated-layout__sidebar">
+      <div className="col-sm-3 authenticated-layout__sidebar">
         <Sidebar />
       </div>
-      <div className="col-md-9 col-md-offset-3 authenticated-layout__main">
+      <div className="col-sm-9 col-sm-offset-3 authenticated-layout__main">
         <Alerter alert={alert} />
         {children}
       </div>
