+++ apps/admin_panel/assets/src/styles/components/_omg_navbar_collapse.scss
@@ -1,0 +1,4 @@
+.omg-navbar-collapse {
+  padding: 0rem;
+  border-top: 0rem;
+}
