+++ apps/admin_panel/assets/src/styles/styles.scss
@@ -22,6 +22,7 @@ $fa-font-path: '~font-awesome/fonts';
 @import 'components/_omg_blue_link.scss';
 @import 'components/_omg_member_item.scss';
 @import 'components/_omg_pager.scss';
+@import 'components/_omg_navbar_collapse.scss';
 
 @import 'containers/_app.scss';
 @import 'containers/_public-layout.scss';
